requirements.txt
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imageio==2.9.0
kiwisolver==1.1.0
matplotlib==3.0.3
networkx==2.4
numpy==1.18.5
opencv-python==4.4.0.42
Pillow==7.2.0
pkg-resources==0.0.0
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.15.0
scipy==1.4.1
six==1.15.0
torch==1.5.1
torchvision==0.6.1
absl-py==0.11.0
